+++ workflow/workflow_apis/src/main.rs
@@ -17,7 +17,7 @@ mod errors;
 mod schema;
 use actix_web::web::Data;
 
-mod test;
+mod tests;
 
 #[actix_web::main]
 async fn main() -> Result<()> {
@@ -39,29 +39,4 @@ async fn main() -> Result<()> {
     .await?;
 
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use actix_web::{http::header::ContentType, test, web, App};
-
-    use super::*;
-
-    #[actix_web::test]
-    async fn test_index_get() {
-        let app = test::init_service(App::new().route("/", web::get().to(index))).await;
-        let req = test::TestRequest::default()
-            .insert_header(ContentType::plaintext())
-            .to_request();
-        let resp = test::call_service(&app, req).await;
-        assert!(resp.status().is_success());
-    }
-
-    #[actix_web::test]
-    async fn test_index_post() {
-        let app = test::init_service(App::new().route("/", web::get().to(index))).await;
-        let req = test::TestRequest::post().uri("/").to_request();
-        let resp = test::call_service(&app, req).await;
-        assert!(resp.status().is_client_error());
-    }
 }
